refactor(ui): generalize GameOverPopup and fix snake game state
Make GameOverPopup a more generic ActionPopup to be reused in other parts of the app.

The new ActionPopup is more flexible and accepts a content widget and a list of actions.

Add an image of a sad Odin to the popup.

Fix a LateInitializationError in the snake game screen by moving the game initialization to initState.