style(ui): replace text buttons with icon buttons on result screen
This change replaces the text-based 'ChibiButton' widgets on the result screen with icon-based 'ChibiButton' widgets.

This modification aligns the result screen’s UI with the design of other components, such as the victory popups, creating a more consistent user experience. The text labels for 'Share' and 'Home' have been replaced by 'share' and 'home' icons, respectively.